Add files via upload
For the Network Security course (part of the Google CyberSecurity certification), I was tasked with writing a security risk assessment report for a fictitious social media organization which had recently suffered a data breach, which could have been caused by a variety of reasons. I was to choose 3 types of network hardening methods out of a small list of many, and determine why and how they would improve network security.